sql: new syntax for dropping field constraints
This patch allows to define the field from which the constraint will be
dropped.

Part of #6986

@TarantoolBot document
Title: DROP CONSTRAINT now able to drop BOX constraints

ALTER TABLE DROP CONSTRAINT can now be used to drop BOX field and tuple
constraints. To drop a constraint, you can use the standard syntax:
```
ALTER TABLE <table name> DROP CONSTRAINT <constraint name>;
```

If there is more than one constraint with the same name, then one of
them will be dropped.

Also, to specify field constraint to drop a new syntax was introduced:
```
ALTER TABLE <table name> DROP CONSTRAINT <field name>.<constraint name>;
```

In this case, the constraint with the given name will be searched among
the constraints of the field. If it is found, it will be dropped.