box: introduce SQL_EXPR to field constraints
After this patch, it will be possible to use the SQL_EXPR functions not
only for tuple constraints, but also for field constraints.

Needed for #6986

@TarantoolBot document
Title: SQL_EXPR functions

Now it is possible to create functions with `SQL_EXPR` language. The
`body` field of the function definition is the SQL expression.
Currently, these functions can only be used as constraints on tuples and
fields, as there is no interface to define argument names and types.