sql: rework parsing of TABLE RENAME statement
This patch reworks ALTER TABLE RENAME statement parsing. Prior to this
patch, VDBE for ALTER TABLE RENAME statement was generated while the
statement was in the process of being parsed. After this patch, it will
be generated only after the parsing has successfully completed.

Part of #3319
Part of #4621
Part of #5485
Part of #8100

NO_DOC=Will be added later
NO_CHANGELOG=Will be added later