sql: remove MEM_Static flag
This patch removes the MEM_Static flag. This flag was used to indicate
that a variable length MEM value will not change during the lifetime of
the MEM, so there is no need to manage the memory where the value is
located. Since this flag was introduced as an counterpart for ephemeral
and dynamic types, it is not really needed.

Part of #6051

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring