sql: move z and n to the union in the struct MEM
Prior to this patch, variable length values were not stored in a union
containing all other values in a struct MEM. This is legacy, as it was
possible to store not one, but two representations of a value this way.
For example, both 1 and "1" for an integer value. Now this is not
possible, so there is no point in storing variable length values not in
a union.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring