test: exclude msgpack module from override cases
The msgpack module can't lack `msgpack.NULL`, because it leads to the
'Please call box.cfg{} first' error at access to `box.NULL`.

An upcoming patch regarding schemas for tarantool's new hierarchical
configuration uses `box.NULL` during tarantool's initialization.

The override feature is tested on a set of modules, which are easy to
replace with an arbitrary table without losing an ability to pass
tarantool's initialization. The msgpack module will not meet the
criteria after the mentioned patch.

The test doesn't attempt to hold a precise list of modules available for
overriding. It rather checks the overall mechanics. So it is OK to
adjust the list.

NO_DOC=testing code change
NO_CHANGELOG=see NO_DOC