conf: make <schema>:merge() schema-aware
The main reason is to distinguish a record from an array. We should
perform a deep merge for a record, but not for an array.

In fact, the previous implementation was buggy not even counting the
point regarding arrays.

```
 1 merge_impl = function(a, b)
 2    if type(a) ~= 'table' then
 3        return b
 4    end
 5    if b == nil then
 6        return a
 7    end
 8    assert(type(b) == 'table')
 9    local res = {}
10    for k, v in pairs(a) do
11        res[k] = merge_impl(v, b[k])
12    end
13    <...>
```

If `v` is non-nil and `b[k]` is `nil` on line 11, then the
`merge_impl()` call will return `nil` anyway due to the if-condition on
line 2.