conf: drop use_defualt flag from <schema>:get()
It is unused. <schema>:apply_default() is more convenient in our case.

Moreover, the use_default flag in :get() was implemented in a bit
strange way. If there is a data from a record, it is returned as is
despite that there may be missed fields that has its defaults.