conf: copy the path in :filter() and :walkthrough()
The `ctx.path` is a mutable table -- it changes during the traversal.

It is usually ok, because :walkthrough() just calls a usr provided
callback. There was an assumption that all the data will be processed in
the callback.

However, it is not the case for :filter(). It accumulates returned data
and returns an iterator on them.