cond: drop <schema object>:walkthrough()
It is not used outside of the schema.lua itself in the implementation of
the <schema object>:filter() function.

The filter functions is easier to implement directly, without extra
abstraction levels.

Found a bug with idempotence of the config applying and fixed. Before
the change `box.NULL` values from `default` annotations were filtered
out by :filter(), so a previously applied value was kept.