test/config: adjust CLI options test
The CLI options test is a unit test in fact and doesn't suppose to test
tarantool startup logic. Let's disable this logic using an override
module generated during the test.

Starting with `--name <...>` and `--config <...>` options is to be
tested as part of the test/config-luatest test suite.

NO_DOC=modification of a testing code
NO_CHANGELOG=see NO_DOC