box: allow SQL_EXPR functions in defaults
After this patch SQL_EXPR functions will be allowed to set as default
values.

Part of #8793

@TarantoolBot document
Title: SQL_EXPR functions as default values

SQL_EXPR functions can now be set as default values. Note that all
arguments to the SQL_EXPR functions will be replaced by the space field
'default' value. This only matters for the SQL_EXPR used as default from
Lua, since in SQL it is not allowed to have arguments in the default
expression.