sql: rework field zSpan of struct SrcList_item
This patch turns the "char *zSpan" field into a "struct Token span"
field. This allows to get the original expression instead of the
unquoted version.

This solution is not the best, since the expression is usually not a
token, but using the “struct ExprSpan” here is quite inconvenient.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring