sql: introduce function default values
This patch replaces the SQL mechanism for setting expression default
values with the BOX mechanism.

Closes #8793

@TarantoolBot document
Title: field default value in SQL

SQL now uses the default mechanism for BOX fields. The main difference
visible to the user is that now all NULL values explicitly inserted into
a field with a default value will be replaced with a default value,
which is different from the original SQL behavior.